Restructure providers
